# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OnlinePort",
        "version": "0.1"
    },
    "paths": {
...
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Show Address",
                "description": "Show Address",
                "operationId": "cb32d4583dda5894217fc7399edcc54d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update Address",
                "description": "Update Address",
                "operationId": "Update Address",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rossi Boutique Hotel & SPA"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "59.9310551"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "30.3338147"
                                    },
                                    "address_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Summary",
                "description": "Addresses",
                "operationId": "Delete Address",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/address-types": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get list of Address Types",
                "operationId": "cb2f56fed79cc263de9e943f4ab9cf25",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Store Address Type",
                "operationId": "f4017f52de134af5733ae71bea585572",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "AddressType1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/address-types/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Show Address Type",
                "description": "Show Address Type",
                "operationId": "f2b8c7382713209a58b31d64ebb7a4a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update Address Type",
                "description": "Update Address Type",
                "operationId": "Update Address Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "AddressTypeTest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Summary",
                "description": "Addresses",
                "operationId": "Delete Address Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}